Fixed compilation errors and warnings appeared after recent changes in the NetSchedule API. JIRA: CXX-2681, CXX-2743, CXX-2836, CXX-2922, CXX-2923
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@53176 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/grid/grid_cli/ns_cmds.cpp
@@ -680,7 +680,7 @@ int CGridCommandLineInterfaceApp::Cmd_SubmitJob()
 
                 printf(format, GetFastLocalTime().
                         AsString(s_NotificationTimestampFormat).c_str(),
-                        submit_job_handler.GetMessage().size(),
+                        (int) submit_job_handler.GetMessage().size(),
                         submit_job_handler.GetMessage().data(),
                         submit_job_handler.GetServerHost().c_str(),
                         submit_job_handler.GetServerPort(),
@@ -894,7 +894,7 @@ int CGridCommandLineInterfaceApp::Cmd_RequestJob()
                     printf("%s \"%.*s\" %s:%u [%s]\n",
                             GetFastLocalTime().AsString(
                                 s_NotificationTimestampFormat).c_str(),
-                            wait_job_handler.GetMessage().size(),
+                            (int) wait_job_handler.GetMessage().size(),
                             wait_job_handler.GetMessage().data(),
                             wait_job_handler.GetServerHost().c_str(),
                             wait_job_handler.GetServerPort(),
